+++ src/controllers/UsersController.php
@@ -1820,31 +1820,36 @@ class UsersController extends Controller
 
         $newPhoto = false;
         $fileLocation = null;
-        $photoName = null;
+        $filename = null;
 
         // Did they upload a new one?
         if ($photo = UploadedFile::getInstanceByName('photo')) {
             $fileLocation = Assets::tempFilePath($photo->getExtension());
             move_uploaded_file($photo->tempName, $fileLocation);
-            $photoName = $photo->name;
+            $filename = $photo->name;
             $newPhoto = true;
-
-        // Did they post a new one?
         } else if (($photo = $this->request->getBodyParam('photo')) && is_array($photo)) {
+            // base64-encoded photo
             $matches = [];
 
-            if (preg_match('/^data:(?<type>[a-z0-9]+\/[a-z0-9\+]+);base64,(?<data>.+)/i', $photo['data'] ?? '', $matches)) {
-                $photoName = empty($photo['filename']) ? null : $photo['filename'];
-                $mimeType = $matches['type'];
-                $data = base64_decode($matches['data']);
-                $extension = FileHelper::getExtensionByMimeType($mimeType);
-                $fileLocation = Assets::tempFilePath($extension);
+            if (preg_match('/^data:((?<type>[a-z0-9]+\/[a-z0-9\+]+);)?base64,(?<data>.+)/i', $photo['data'] ?? '', $matches)) {
+                $filename = $photo['filename'] ?? null;
+                $extension = $filename ? pathinfo($filename, PATHINFO_EXTENSION) : null;
 
-                if (empty($extension)) {
-                    Craft::warning("Could not resolve extension for mimetype $mimeType");
+                if (!$extension && !empty($matches['type'])) {
+                    try {
+                        $extension = FileHelper::getExtensionByMimeType($matches['type']);
+                    } catch (InvalidArgumentException $e) {
+                    }
+                }
+
+                if (!$extension) {
+                    Craft::warning('Could not determine file extension for user photo.', __METHOD__);
                     return;
                 }
 
+                $fileLocation = Assets::tempFilePath($extension);
+                $data = base64_decode($matches['data']);
                 FileHelper::writeToFile($fileLocation, $data);
                 $newPhoto = true;
             }
@@ -1852,7 +1857,7 @@ class UsersController extends Controller
 
         if ($newPhoto) {
             try {
-                $users->saveUserPhoto($fileLocation, $user, $photoName);
+                $users->saveUserPhoto($fileLocation, $user, $filename);
             } catch (\Throwable $e) {
                 if (file_exists($fileLocation)) {
                     FileHelper::unlink($fileLocation);